Updated version info
Game is finished.

This is the game in it's first version. I need to add animations and transitions, but for the most part the game is done.

One problem:

When you mash a button it skips like crazy so I gotta come up with a way to where the player can mash a button but the game goes at the pace you normally would go waiting for each animation to finish and sytem prompt.